+++ lib/authem/errors/unknown_role.rb
@@ -1,0 +1,8 @@
+module Authem
+  class UnknownRoleError < StandardError
+    def initialize(record)
+      message = "Unknown authem role: #{record.inspect}"
+      super message
+    end
+  end
+end