+++ lib/generators/authem/session/templates/create_sessions.rb
@@ -1,0 +1,15 @@
+class CreateAuthemSessions < ActiveRecord::Migration
+  def change
+    create_table :authem_sessions do |t|
+      t.string     :role,       null: false
+      t.references :subject,    null: false, polymorphic: true
+      t.string     :token,      null: false, limit: 60
+      t.datetime   :expires_at, null: false
+      t.integer    :ttl,        null: false
+      t.timestamps
+    end
+
+    add_index :authem_sessions, %i[expires_at token], unique: true
+    add_index :authem_sessions, %i[expires_at subject_type subject_id]
+  end
+end