+++ spec/spec_helper.rb
@@ -1,11 +1,4 @@
-require 'bundler'
+require "bundler/setup"
+require "authem"
 
-Bundler.require(:default, :development)
-
-# Requires supporting files with custom matchers and macros, etc,
-# in ./support/ and its subdirectories.
-Dir[File.expand_path(File.join(File.dirname(__FILE__),'support','**','*.rb'))].each {|f| require f}
-
-RSpec.configure do |config|
-  config.mock_with :rspec
-end
+Dir[File.expand_path("../support/**/*.rb", __FILE__)].each(&method(:require))