+++ spec/support/time.rb
@@ -1,0 +1,1 @@
+Time.zone_default = ActiveSupport::TimeZone["UTC"]